Fix install.sh for neovim

+++ install.sh
@@ -20,22 +20,29 @@ for file in $files; do
         echo "Symlinking $file..."
         ln -s $PWD/$file ~/.$file
     fi
+done;
 
-    # Custom links for neovim
-    if type nvim > /dev/null; then
-        ln -s ~./vim ~/.config/nvim
+# Custom links for neovim
+if type nvim > /dev/null; then
+    if [ ! -d ~/.config/nvim ]; then
+        echo "Symlinking ~/.config/nvim..."
+        ln -s ~/.vim ~/.config/nvim
+    fi
+
+    if [ ! -f ~/.config/nvim/init.vim ]; then
+        echo "Symlinking ~/.config/nvim/init.vim..."
         ln -s ~/.vimrc ~/.config/nvim/init.vim
     fi
-done;
+fi
 
 # Install vim plugins
 echo "Installing vim plugins..."
-vim -c PlugInstall -c q
+vim -c PlugInstall -c q -c q
 
 # Install neovim plugins
 if type nvim > /dev/null; then
     echo "Installing neovim plugins..."
-    nvim -c PlugInstall -c q
+    nvim -c PlugInstall -c q -c q
 fi
 
 # Ask for user and email for gitconfig
